+++ KSP-Backend/Pipfile
@@ -1,0 +1,64 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+admin = "==0.0.1"
+appdirs = "==1.4.4"
+asgiref = "==3.3.4"
+certifi = "==2020.12.5"
+chardet = "==4.0.0"
+click = "==7.1.2"
+cycler = "==0.10.0"
+distlib = "==0.3.1"
+dj-database-url = "==0.5.0"
+dj-static = "==0.0.6"
+django-autoslug = "==1.9.8"
+django-comments-xtd = "==2.9.0"
+django-contrib-comments = "==2.1.0"
+django-jquery = "==3.1.0"
+django-js-asset = "==1.2.2"
+django-summernote = "==0.8.11.6"
+django-vote = "==2.2.0"
+djangorestframework = "==3.12.4"
+docutils = "==0.17.1"
+filelock = "==3.0.12"
+gunicorn = "==20.1.0"
+idna = "==2.10"
+joblib = "==1.0.1"
+kiwisolver = "==1.3.1"
+matplotlib = "==3.4.1"
+mysql-connector = "==2.2.9"
+mysqlclient = "==2.0.3"
+nltk = "==3.6.2"
+numpy = "==1.20.2"
+oauthlib = "==3.1.0"
+psycopg2 = "==2.8.6"
+pyparsing = "==2.4.7"
+python-dateutil = "==2.8.1"
+python-rake = "==1.5.0"
+pytz = "==2021.1"
+regex = "==2021.4.4"
+requests = "==2.25.1"
+requests-oauthlib = "==1.3.0"
+six = "==1.15.0"
+sqlparse = "==0.4.1"
+static3 = "==0.7.0"
+textblob = "==0.15.3"
+tqdm = "==4.60.0"
+tweepy = "==3.10.0"
+urllib3 = "==1.26.4"
+virtualenv = "==20.4.2"
+virtualenvwrapper-win = "==1.2.6"
+whitenoise = "==5.2.0"
+Django = "==3.2"
+Pillow = "==8.2.0"
+PySocks = "==1.7.1"
+vaderSentiment = "==3.3.2"
+django-cors-headers = "*"
+
+[dev-packages]
+
+[requires]
+python_version = "3.9"
